작성자 추가를 위한 스키마 변경

+++ src/main/java/com/back/domain/member/member/service/MemberService.java
@@ -52,6 +52,7 @@ public class MemberService {
     }
 
     public int deleteByIds(List<Integer> ids) {
+        if (ids.isEmpty()) return 0;
         return memberRepository.deleteByIds(ids);
     }
 }
